+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * **Raphael Pour**
     * cmd/root: add go version and last commit sha to the --version build information
     * cmd/render: support opengraph tags for rich link preview
+    * metadata: pretty render initial metadata
 
 *Not released yet*