+++ CHANGELOG.md
@@ -3,6 +3,7 @@
     * cmd/root: add go version and last commit sha to the --version build information
     * cmd/render: support opengraph tags for rich link preview
     * metadata: pretty render initial metadata
+    * cmd/render, config: add chill-files list containing arbitrary files that just get copied to the root of the output dir
 
 *Not released yet*