+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * metadata: pretty render initial metadata
     * cmd/render, config: add chill-files list containing arbitrary files that just get copied to the root of the output dir
     * cmd: add admin/rewrite-metadata command to pretty print all old metadata files
+    * cmd/list: render table via tablewriter
 
 *Not released yet*